fix: divide router declaration and routes implementing
